fix: Seperate out loading of config to common
