+++ src/internal/common/ReadMe.md
@@ -2,3 +2,6 @@
 Defines common utilities for ui and file operations packge
 everyone can use common packge, but common package should not have any dependency for
 any other package
+
+# Dependencies
+- src/config package
