+++ src/internal/common/prerendered_variable.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"github.com/charmbracelet/lipgloss"
