+++ src/internal/common/utils/ReadMe.md
@@ -1,0 +1,2 @@
+# utils package
+Independent utilities with zero dependency with other packages
